Fix review findings: StreamConfig, header split, type tags
- Fix StreamConfig{nullptr,false} bug in depthwise impl by using
  handle.GetStream() and switching to MakeArgumentPointer/MakeInvokerPointer
  API to match all other CK impl files
- Move CK-dependent templates (FindConvPtrByID, FillValidKernelsIDs,
  DebugPrintCKArgPtrs) from common header to CK-side header since they
  call CK device methods and are only used by ck_impl/ files
- Remove unused bias_lens/bias_strides members from depthwise CKArgs
- Share HalfTag/BFloat16Tag/TF32Tag in common header instead of
  duplicating in 3 anonymous namespaces across 3D solver files
- Add data_type/problem null checks to get_workspace_size for consistency
- Update test comment to mention CK dynamic library dependency
