+++ projects/miopen/test/gtest/unit_conv_solver_ConvCkGroupedConvFwd.cpp
@@ -117,7 +117,7 @@ auto GetConvFullTestCases(miopenDataType_t datatype)
 
 auto GetTestParams(miopenDataType_t datatype)
 {
-// Solution is only applicable to devices that default to 64-lane wavefronts
+// Solution requires 64-lane wavefronts and depends on the CK dynamic library
 #if MIOPEN_BACKEND_HIP && MIOPEN_USE_COMPOSABLEKERNEL
     Gpu supportedDevices = Gpu::gfx908 | Gpu::gfx90A | Gpu::gfx94X | Gpu::gfx950;
 #else
